add aiter rope hip forward func
